# Conflicts:
#	src/main/resources/biz_bg.properties
#	src/main/resources/biz_cs.properties
#	src/main/resources/biz_da.properties
#	src/main/resources/biz_en.properties
#	src/main/resources/biz_es.properties
#	src/main/resources/biz_fi.properties
#	src/main/resources/biz_fr.properties
#	src/main/resources/biz_hu.properties
#	src/main/resources/biz_it.properties
#	src/main/resources/biz_nl.properties
#	src/main/resources/biz_pl.properties
#	src/main/resources/biz_pt.properties
#	src/main/resources/biz_ro.properties
#	src/main/resources/biz_ru.properties
#	src/main/resources/biz_sk.properties
#	src/main/resources/biz_sv.properties

+++ src/main/java/sirius/biz/cluster/NeighborhoodWatch.java
@@ -330,7 +330,7 @@ public class NeighborhoodWatch implements Orchestration, Initializable, Intercon
                         db.setnx(syncName + EXECUTION_TIMESTAMP_SUFFIX, String.valueOf(System.currentTimeMillis()));
                 if (update == 1L) {
                     db.expire(syncName + EXECUTION_TIMESTAMP_SUFFIX,
-                              (int) TimeUnit.HOURS.toSeconds(MIN_WAIT_DAILY_TASK_HOURS));
+                              (long) TimeUnit.HOURS.toSeconds(MIN_WAIT_DAILY_TASK_HOURS));
                     return true;
                 } else {
                     return false;
